fixed (i hope) keyscan issue with root permissions

+++ scripts/server_scripts/master_scripts/reset_all.sh
@@ -1,26 +1,33 @@
 #!/bin/bash
 
-# TODO: Running the script with sudo permissions uses /root/.ssh instead of using /home/user/.ssh. Find a way to fix this.
-# clear all non hidden files in mysql server
+if [ "$EUID" -eq 0 ] # Root has $EUID = 0
+  then echo "Please run this script as non-root (no sudo)."
+  exit
+fi
+
+# clear mysql server
 source ../config/config_mysql.sh
-su -c "ssh-keygen -R $server_ip" $USER 
-su -c "ssh-keyscan -H $server_ip >> ~/.ssh/known_hosts" $USER
+echo "clearing mysql server"
+ssh-keygen -R $server_ip
+ssh-keyscan -t ecdsa -H $server_ip >> ~/.ssh/known_hosts
 
 ssh -i ~/.ssh/$public_key $username@$server_ip "sudo rm -rf /home/ubuntu/*"
-su -c "ssh-keygen -R $server_ip" $USER
+ssh-keygen -R $server_ip # remove key at the end cleanup
 
-# clear all non hidden files in mongodb server
-# source ../config/config_mongodb.sh
-# sudo -u $USER ssh-keygen -R $server_ip
-# sudo -u $USER ssh-keyscan -H $server_ip >> ~/.ssh/known_hosts
+# clear mongodb server
+source ../config/config_mongodb.sh
+echo "clearning mongodb server"
+ssh-keygen -R $server_ip
+ssh-keyscan -t ecdsa -H $server_ip >> ~/.ssh/known_hosts
 
-# ssh -i ~/.ssh/$public_key $username@$server_ip "sudo rm -rf /home/ubuntu/*"
-# sudo -u $USER ssh-keygen -R $server_ip
+ssh -i ~/.ssh/$public_key $username@$server_ip "sudo rm -rf /home/ubuntu/*"
+ssh-keygen -R $server_ip # remove key at the end cleanup
 
-# clear all non hidden files in flask server
-# source ../config/flask.sh
-# ssh-keygen -R $server_ip
-# ssh-keyscan -H $server_ip >> /root/.ssh/known_hosts
-# ssh-keygen -R $server_ip
+# clear flask server
+echo "clearing flask server"
+source ../config/flask.sh
+ssh-keygen -R $server_ip
+ssh-keyscan -t ecdsa -H $server_ip >> ~/.ssh/known_hosts
 
-# ssh -i ~/.ssh/$public_key $username@$server_ip "sudo rm -rf /home/ubuntu/*"
+ssh -i ~/.ssh/$public_key $username@$server_ip "sudo rm -rf /home/ubuntu/*"
+ssh-keygen -R $server_ip # remove key at the end cleanup
